Split runner project creation tests
