+++ crates/webcodex-runner/src/main_tests.rs
@@ -124,6 +124,8 @@ mod http_recovery;
 mod managed_temporary_projects;
 #[path = "main_tests/profile_process_lifecycle.rs"]
 mod profile_process_lifecycle;
+#[path = "main_tests/project_creation.rs"]
+mod project_creation;
 #[path = "main_tests/project_durability.rs"]
 mod project_durability;
 #[path = "main_tests/project_policy.rs"]
@@ -1723,166 +1725,6 @@ fn project_error_value(result: CommandResult) -> serde_json::Value {
     );
     assert!(result.error.is_none(), "unexpected raw error: {:?}", result);
     serde_json::from_str(result.stdout.as_deref().expect("error json")).unwrap()
-}
-
-#[test]
-fn create_project_basic_creates_readme_and_gitignore() {
-    let tmp = tempfile::tempdir().unwrap();
-    let project_dir = tmp.path().join("new-project");
-    let projects_dir = tmp.path().join("projects.d");
-    let policy = project_policy(tmp.path());
-    let req = project_request(
-        "create_project",
-        serde_json::json!({
-            "id": "basic",
-            "name": "Basic",
-            "path": project_dir.to_string_lossy(),
-            "description": "Basic template",
-            "template": "basic"
-        }),
-    );
-
-    let value = project_ok(handle_project_op(&policy, &projects_dir, &req));
-    assert_eq!(value["created_directory"], true);
-    assert!(project_dir.join("README.md").exists());
-    assert!(project_dir.join(".gitignore").exists());
-    assert!(std::fs::read_to_string(project_dir.join("README.md"))
-        .unwrap()
-        .contains("Basic template"));
-}
-
-#[test]
-fn create_project_rejects_existing_non_empty_directory() {
-    let tmp = tempfile::tempdir().unwrap();
-    let project_dir = tmp.path().join("existing");
-    let projects_dir = tmp.path().join("projects.d");
-    std::fs::create_dir(&project_dir).unwrap();
-    let keep = project_dir.join("keep.txt");
-    std::fs::write(&keep, "keep").unwrap();
-    let policy = project_policy(tmp.path());
-    let req = project_request(
-        "create_project",
-        serde_json::json!({
-            "id": "existing",
-            "name": "Existing",
-            "path": project_dir.to_string_lossy(),
-            "template": "basic",
-            "allow_existing_empty": true
-        }),
-    );
-
-    let err = project_err(handle_project_op(&policy, &projects_dir, &req));
-    assert_eq!(err, "path_not_empty");
-    assert_eq!(std::fs::read_to_string(keep).unwrap(), "keep");
-}
-
-#[test]
-fn create_project_rejects_unknown_template() {
-    let tmp = tempfile::tempdir().unwrap();
-    let project_dir = tmp.path().join("new-project");
-    let projects_dir = tmp.path().join("projects.d");
-    let policy = project_policy(tmp.path());
-    let req = project_request(
-        "create_project",
-        serde_json::json!({
-            "id": "badtemplate",
-            "name": "Bad Template",
-            "path": project_dir.to_string_lossy(),
-            "template": "cargo"
-        }),
-    );
-
-    let err = project_err(handle_project_op(&policy, &projects_dir, &req));
-    assert_eq!(err, "invalid_request");
-    assert!(!project_dir.exists());
-}
-
-#[test]
-fn create_project_created_config_and_overwritten_semantics_are_accurate() {
-    let tmp = tempfile::tempdir().unwrap();
-    let project_dir = tmp.path().join("empty-project");
-    let projects_dir = tmp.path().join("projects.d");
-    let policy = project_policy(tmp.path());
-    let payload = |overwrite| {
-        serde_json::json!({
-            "id": "empty",
-            "name": "Empty",
-            "path": project_dir.to_string_lossy(),
-            "template": "empty",
-            "allow_existing_empty": true,
-            "overwrite": overwrite
-        })
-    };
-
-    let first = project_ok(handle_project_op(
-        &policy,
-        &projects_dir,
-        &project_request("create_project", payload(false)),
-    ));
-    assert_eq!(first["created_directory"], true);
-    assert_eq!(first["created_config"], true);
-    assert_eq!(first["overwritten"], false);
-
-    let second = project_ok(handle_project_op(
-        &policy,
-        &projects_dir,
-        &project_request("create_project", payload(true)),
-    ));
-    assert_eq!(second["created_directory"], false);
-    assert_eq!(second["created_config"], false);
-    assert_eq!(second["overwritten"], true);
-}
-
-#[test]
-fn create_project_cleanup_removes_only_files_created_on_failure() {
-    let tmp = tempfile::tempdir().unwrap();
-    let project_dir = tmp.path().join("existing-empty");
-    std::fs::create_dir(&project_dir).unwrap();
-    let projects_dir_file = tmp.path().join("projects.d-is-file");
-    std::fs::write(&projects_dir_file, "not a dir").unwrap();
-    let policy = project_policy(tmp.path());
-    let req = project_request(
-        "create_project",
-        serde_json::json!({
-            "id": "cleanup",
-            "name": "Cleanup",
-            "path": project_dir.to_string_lossy(),
-            "template": "basic",
-            "allow_existing_empty": true
-        }),
-    );
-
-    let err = project_err(handle_project_op(&policy, &projects_dir_file, &req));
-    assert_eq!(err, "operation_failed");
-    assert!(project_dir.exists());
-    assert!(!project_dir.join("README.md").exists());
-    assert!(!project_dir.join(".gitignore").exists());
-}
-
-#[test]
-fn create_project_does_not_delete_pre_existing_files() {
-    let tmp = tempfile::tempdir().unwrap();
-    let project_dir = tmp.path().join("existing");
-    std::fs::create_dir(&project_dir).unwrap();
-    let pre_existing = project_dir.join("pre-existing.txt");
-    std::fs::write(&pre_existing, "original").unwrap();
-    let projects_dir_file = tmp.path().join("projects.d-is-file");
-    std::fs::write(&projects_dir_file, "not a dir").unwrap();
-    let policy = project_policy(tmp.path());
-    let req = project_request(
-        "create_project",
-        serde_json::json!({
-            "id": "keep",
-            "name": "Keep",
-            "path": project_dir.to_string_lossy(),
-            "template": "basic",
-            "allow_existing_empty": true
-        }),
-    );
-
-    let err = project_err(handle_project_op(&policy, &projects_dir_file, &req));
-    assert_eq!(err, "path_not_empty");
-    assert_eq!(std::fs::read_to_string(pre_existing).unwrap(), "original");
 }
 
 #[test]
